Revert use of contains()
This change could result in problems for other packages that create fastmap
objects at package build time. If pkgA is rebuilt on CRAN with this new
implementation of has() and then a user installs it, but does not update
fastmap, then it will result in an error when they call has(), because the
C function does not exist in the old version of fastmap.